requirements.txt
PyYAML
send2trash
croniter